// web/src/lib/schemas.ts
export const AccountBase = z.object({
  name: z.string(),
  type: z.enum(["savings", "checking", "creditCard"]),
  balance: z.number(),
  userId: IdSchema,
});

export const AccountSchema = addIdAndDatetime(AccountBase);
export type Account = z.infer<typeof AccountSchema>;